2 New examples, changed iMouse and iResolution to float vectors to fix bug

+++ Wallpaper Engine/default.frag
@@ -9,7 +9,7 @@
 #version 330 core
 
 uniform float iTime;
-uniform ivec2 iResolution;
+uniform vec2 iResolution;
 uniform float timeMultiplier;
 
 uniform vec4 color;
